backend/requirements.txt
# Agent framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
sse-starlette>=2.1.0

# Database
neo4j>=5.27.0

# AWS (mocked in demo, real in prod)
boto3>=1.35.0

# CopilotKit backend SDK
copilotkit>=0.1.39

# Data validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# HTTP client (for TestSprite + Datadog REST fallback)
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.4.0
